refactor(backend): extract query builder and filter helpers to query-builder.ts
Move all filter/sort query helpers from backend.ts to a new src/backend/query-builder.ts module: sorting helpers (exampleFileDTO, isFileDTOPropString, PaginationOptions, applyPagination, getOrderColumnExpression), filter helpers (ConditionWithConjunction, applyFileFilters, expressionFromNode, expressionFromCriteria, applyStringCondition, applyNumberCondition, applyDateCondition, applyTagArrayCondition, applyExtraPropertyCondition), and semantic video helpers (computeSampleTimestamps, meanPoolEmbeddings) as a temporary home pending a future SemanticRepository.
